requirements.txt
streamlit>=1.28
folium>=0.15
streamlit-folium>=0.18.0
fastapi>=0.109
uvicorn[standard]>=0.27
pydantic>=2
pydantic-settings>=2
sqlalchemy>=2
python-dotenv>=1.0
requests>=2.31
httpx>=0.27
google-generativeai>=0.8.0
osmnx
networkx
numpy>=1.24
